Original change: https://android-review.googlesource.com/c/platform/system/memory/libmeminfo/+/2099487

+++ libdmabufinfo/include/dmabufinfo/dmabuf_sysfs_stats.h
@@ -30,7 +30,7 @@ namespace dmabufinfo {
  * @size: Size of the buffer.
  */
 struct DmabufInfo {
-    unsigned int inode;
+    unsigned long inode;
     std::string exp_name;
     unsigned int size;
 };
